requirements.txt
selenium
undetected_chromedriver
regex
python-dotenv
flask